Rename to [X-chromosomal ancestors of <person>]
